+++ core/core.go
@@ -295,53 +295,6 @@ func ParseNested(input string) []string {
 	return result
 }
 
-func ApplyFilter(builder QueryBuilder, filter Filter, fieldExprApplier FieldExprApplier) QueryBuilder {
-
-	// Campos
-	for field, expr := range filter.Fields {
-		// builder = applyFieldExpr(builder, field, expr)
-		builder = fieldExprApplier(builder, field, expr)
-	}
-
-	// AND
-	for _, andItemFilter := range filter.And {
-		subBuilder := builder.Clone()
-		ApplyFilter(subBuilder, andItemFilter, fieldExprApplier)
-		builder = builder.And(subBuilder)
-	}
-
-	// OR
-	for _, orItemFilter := range filter.Or {
-		subBuilder := builder.Clone()
-		ApplyFilter(subBuilder, orItemFilter, fieldExprApplier)
-		builder = builder.Or(subBuilder)
-	}
-
-	// NOT
-	if filter.Not != nil {
-		subBuilder := builder.Clone()
-		ApplyFilter(subBuilder, *filter.Not, fieldExprApplier)
-		builder = builder.Not(subBuilder)
-	}
-
-	return builder
-}
-
-func CastIfJSONB(sqlField string, isJSONB bool, value interface{}) string {
-	if !isJSONB {
-		return sqlField
-	}
-
-	switch value.(type) {
-	case int, int32, int64, float32, float64:
-		return sqlField + "::numeric"
-	case bool:
-		return sqlField + "::boolean"
-	default:
-		return sqlField
-	}
-}
-
 func ParseNestedTree(input string) []*NestedNode {
 	input = strings.TrimSpace(input)