+++ gorm/gorm.go
@@ -9,6 +9,7 @@ import (
 	"strings"
 
 	fwork_server_orm "github.com/joabssilveira/GoQLite/core"
+	fwork_server_orm_implementation "github.com/joabssilveira/GoQLite/implementation"
 	"gorm.io/gorm"
 	"gorm.io/gorm/schema"
 )
@@ -66,10 +67,10 @@ func (g *GormQueryBuilder) Clone() fwork_server_orm.QueryBuilder {
 	}
 }
 
-func ApplyQuery(builder *GormQueryBuilder, payload fwork_server_orm.QueryPayload) *GormQueryBuilder {
+func ApplyQuery(builder *GormQueryBuilder, payload fwork_server_orm.QueryPayload, dbUtils fwork_server_orm_implementation.DbUtils) *GormQueryBuilder {
 	// WHERE
 	ApplyJoinsFromFilter(builder.Db, builder.Db.Statement.Model, payload.Where)
-	builder = fwork_server_orm.ApplyFilter(builder, payload.Where, applyFieldExpr).(*GormQueryBuilder)
+	builder = fwork_server_orm_implementation.ApplyFilter(builder, payload.Where, applyFieldExpr, dbUtils).(*GormQueryBuilder)
 
 	// SELECT
 	if len(payload.Select) > 0 {
@@ -118,7 +119,7 @@ func ApplyQuery(builder *GormQueryBuilder, payload fwork_server_orm.QueryPayload
 	if payload.Nested != "" {
 		tree := fwork_server_orm.ParseNestedTree(payload.Nested)
 		for _, node := range tree {
-			applyNestedNode(builder.Db, builder.Db.Statement.Model, node, "")
+			applyNestedNode(builder.Db, builder.Db.Statement.Model, node, "", dbUtils)
 		}
 	}
 
@@ -133,7 +134,7 @@ func toGormRelationPath(path string) string {
 	return strings.Join(parts, ".")
 }
 
-func applyNestedNode(db *gorm.DB, parentModel any, node *fwork_server_orm.NestedNode, prefix string) {
+func applyNestedNode(db *gorm.DB, parentModel any, node *fwork_server_orm.NestedNode, prefix string, dbUtils fwork_server_orm_implementation.DbUtils) {
 	// resolve nome real da relação no struct
 	gormName := toGormRelationPath(node.Name)
 
@@ -161,7 +162,7 @@ func applyNestedNode(db *gorm.DB, parentModel any, node *fwork_server_orm.Nested
 			}
 
 			sub := NewGormQueryBuilder(tx)
-			sub = ApplyQuery(sub, *node.Query)
+			sub = ApplyQuery(sub, *node.Query, dbUtils)
 			return sub.Db
 		}
 		return tx
@@ -172,7 +173,7 @@ func applyNestedNode(db *gorm.DB, parentModel any, node *fwork_server_orm.Nested
 
 	// recursão
 	for _, child := range node.Childs {
-		applyNestedNode(db, childModel, child, gormPath)
+		applyNestedNode(db, childModel, child, gormPath, dbUtils)
 	}
 }
 
@@ -303,7 +304,24 @@ func applyRelationJoinWithParentAlias(
 	}
 }
 
-func applyFieldExpr(builder fwork_server_orm.QueryBuilder, field string, expr fwork_server_orm.FieldExpr) fwork_server_orm.QueryBuilder {
+// func getUnaccent(s string) string {
+// 	t := norm.NFD.String(s)
+
+// 	var b strings.Builder
+// 	b.Grow(len(t))
+
+// 	for _, r := range t {
+// 		// Remove marcas diacríticas (acentos)
+// 		if unicode.Is(unicode.Mn, r) {
+// 			continue
+// 		}
+// 		b.WriteRune(r)
+// 	}
+
+// 	return norm.NFC.String(b.String())
+// }
+
+func applyFieldExpr(builder fwork_server_orm.QueryBuilder, field string, expr fwork_server_orm.FieldExpr, dbUtils fwork_server_orm_implementation.DbUtils) fwork_server_orm.QueryBuilder {
 	gormBuilder, ok := builder.(*GormQueryBuilder)
 	if !ok {
 		return builder
@@ -448,73 +466,75 @@ func applyFieldExpr(builder fwork_server_orm.QueryBuilder, field string, expr fw
 	// Operadores
 	// =========================
 
-	if expr.Eq != nil {
-		builder = builder.Where(sqlField+" = ?", expr.Eq)
-	}
+	// if expr.Eq != nil {
+	// 	builder = builder.Where(sqlField+" = ?", expr.Eq)
+	// }
 
-	if expr.Ne != nil {
-		builder = builder.Where(sqlField+" <> ?", expr.Ne)
-	}
+	// if expr.Ne != nil {
+	// 	builder = builder.Where(sqlField+" <> ?", expr.Ne)
+	// }
 
-	if expr.Gt != nil {
-		builder = builder.Where(fwork_server_orm.CastIfJSONB(sqlField, isJSONB, expr.Gt)+" > ?", expr.Gt)
-	}
+	// if expr.Gt != nil {
+	// 	builder = builder.Where(fwork_server_orm.CastIfJSONB(sqlField, isJSONB, expr.Gt)+" > ?", expr.Gt)
+	// }
 
-	if expr.Gte != nil {
-		builder = builder.Where(fwork_server_orm.CastIfJSONB(sqlField, isJSONB, expr.Gte)+" >= ?", expr.Gte)
-	}
+	// if expr.Gte != nil {
+	// 	builder = builder.Where(fwork_server_orm.CastIfJSONB(sqlField, isJSONB, expr.Gte)+" >= ?", expr.Gte)
+	// }
 
-	if expr.Lt != nil {
-		builder = builder.Where(fwork_server_orm.CastIfJSONB(sqlField, isJSONB, expr.Lt)+" < ?", expr.Lt)
-	}
+	// if expr.Lt != nil {
+	// 	builder = builder.Where(fwork_server_orm.CastIfJSONB(sqlField, isJSONB, expr.Lt)+" < ?", expr.Lt)
+	// }
 
-	if expr.Lte != nil {
-		builder = builder.Where(fwork_server_orm.CastIfJSONB(sqlField, isJSONB, expr.Lte)+" <= ?", expr.Lte)
-	}
+	// if expr.Lte != nil {
+	// 	builder = builder.Where(fwork_server_orm.CastIfJSONB(sqlField, isJSONB, expr.Lte)+" <= ?", expr.Lte)
+	// }
 
-	if len(expr.In) > 0 {
-		builder = builder.Where(sqlField+" IN ?", expr.In)
-	}
+	// if len(expr.In) > 0 {
+	// 	builder = builder.Where(sqlField+" IN ?", expr.In)
+	// }
 
-	if len(expr.Nin) > 0 {
-		builder = builder.Where(sqlField+" NOT IN ?", expr.Nin)
-	}
+	// if len(expr.Nin) > 0 {
+	// 	builder = builder.Where(sqlField+" NOT IN ?", expr.Nin)
+	// }
 
-	if expr.Like != "" {
-		builder = builder.Where(sqlField+" LIKE ?", "%"+expr.Like+"%")
-	}
+	// if expr.Like != "" {
+	// 	builder = builder.Where(sqlField+" LIKE ?", "%"+expr.Like+"%")
+	// }
 
-	if expr.ILike != "" {
-		builder = builder.Where(sqlField+" ILIKE ?", "%"+expr.ILike+"%")
-	}
+	// if expr.ILike != "" {
+	// 	builder = builder.Where(sqlField+" ILIKE ?", "%"+expr.ILike+"%")
+	// }
 
-	if len(expr.Between) == 2 {
-		builder = builder.Where(
-			fwork_server_orm.CastIfJSONB(sqlField, isJSONB, expr.Between[0])+" BETWEEN ? AND ?",
-			expr.Between[0],
-			expr.Between[1],
-		)
-	}
+	// if len(expr.Between) == 2 {
+	// 	builder = builder.Where(
+	// 		fwork_server_orm.CastIfJSONB(sqlField, isJSONB, expr.Between[0])+" BETWEEN ? AND ?",
+	// 		expr.Between[0],
+	// 		expr.Between[1],
+	// 	)
+	// }
 
-	if expr.Exists != nil {
-		if *expr.Exists {
-			builder = builder.Where(sqlField + " IS NOT NULL")
-		} else {
-			builder = builder.Where(sqlField + " IS NULL")
-		}
-	}
+	// if expr.Exists != nil {
+	// 	if *expr.Exists {
+	// 		builder = builder.Where(sqlField + " IS NOT NULL")
+	// 	} else {
+	// 		builder = builder.Where(sqlField + " IS NULL")
+	// 	}
+	// }
 
-	if expr.IsNull != nil {
-		if *expr.IsNull {
-			builder = builder.Where(sqlField + " IS NULL")
-		} else {
-			builder = builder.Where(sqlField + " IS NOT NULL")
-		}
-	}
+	// if expr.IsNull != nil {
+	// 	if *expr.IsNull {
+	// 		builder = builder.Where(sqlField + " IS NULL")
+	// 	} else {
+	// 		builder = builder.Where(sqlField + " IS NOT NULL")
+	// 	}
+	// }
 
-	if expr.Op != nil {
-		builder = builder.Where(sqlField+" "+expr.Op.Op+" ?", expr.Op.Value)
-	}
+	// if expr.Op != nil {
+	// 	builder = builder.Where(sqlField+" "+expr.Op.Op+" ?", expr.Op.Value)
+	// }
+
+	builder = dbUtils.GetFieldExpr(builder, expr, sqlField, isJSONB)
 
 	return builder
 }
@@ -523,7 +543,7 @@ func quoteIdent(s string) string {
 	return `"` + s + `"`
 }
 
-func GormGetList[T any](db *gorm.DB, payload fwork_server_orm.QueryPayload) (fwork_server_orm.GetListData[T], error) {
+func GormGetList[T any](db *gorm.DB, payload fwork_server_orm.QueryPayload, dbUtils fwork_server_orm_implementation.DbUtils) (fwork_server_orm.GetListData[T], error) {
 	fwork_server_orm.ApplyPagination(&payload)
 
 	// =========================
@@ -549,7 +569,7 @@ func GormGetList[T any](db *gorm.DB, payload fwork_server_orm.QueryPayload) (fwo
 	var list []T
 
 	dataBuilder := NewGormQueryBuilder(db.Model(new(T)))
-	dataBuilder = ApplyQuery(dataBuilder, payload)
+	dataBuilder = ApplyQuery(dataBuilder, payload, dbUtils)
 
 	if err := dataBuilder.Db.Find(&list).Error; err != nil {
 		return fwork_server_orm.GetListData[T]{}, err
@@ -565,7 +585,7 @@ func GormGetList[T any](db *gorm.DB, payload fwork_server_orm.QueryPayload) (fwo
 	}, nil
 }
 
-func GormGetListHttp[T any](db *gorm.DB, r *http.Request, additionalWhere fwork_server_orm.Filter) (fwork_server_orm.GetListData[T], error) {
+func GormGetListHttp[T any](db *gorm.DB, r *http.Request, additionalWhere fwork_server_orm.Filter, dbUtils fwork_server_orm_implementation.DbUtils) (fwork_server_orm.GetListData[T], error) {
 	var payload fwork_server_orm.QueryPayload
 
 	// where
@@ -625,7 +645,7 @@ func GormGetListHttp[T any](db *gorm.DB, r *http.Request, additionalWhere fwork_
 	// It already exists in GormGetList.
 	// fwork_server_orm.ApplyPagination(&payload)
 
-	return GormGetList[T](db, payload)
+	return GormGetList[T](db, payload, dbUtils)
 }
 
 func ApplyJoinsFromFilter(